Add tap_tester base suite and move mocked integration tests (#55)
* Add tap_tester base suite and move mocked integration tests to separate folder

* Increase test coverage and fix copilot comments
